Refactored different Player types into different modules
